requirements.txt
# Core 12thhaus and LangChain
langgraph>=0.5.2
langsmith>=0.4.4
langchain-anthropic>=0.3.17
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-core>=0.1.0

# Environment and Configuration
python-dotenv>=1.1.1
pydantic>=2.7.4
pydantic-settings>=2.0.0

# Database and Storage
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
redis>=5.0.0

# Supabase Integration
supabase>=2.0.0
postgrest>=0.13.0

# Authentication (Logto Integration)
logto>=0.2.1
python-jose[cryptography]>=3.3.0
flask-cors>=4.0.0
pyjwt>=2.8.0
cryptography>=41.0.0

# HTTP and API
httpx>=0.25.0
requests>=2.31.0
fastapi>=0.104.0
uvicorn>=0.24.0
# Web UI (Streamlit)
streamlit>=1.28.0
streamlit-chat>=0.1.1
plotly>=5.15.0
pandas>=2.0.0

# GitHub Integration
pygithub>=1.59.0
gitpython>=3.1.0

# File Processing
PyYAML>=6.0.1
python-multipart>=0.0.6
aiofiles>=23.0.0

# Code Generation and Templates
jinja2>=3.1.0
autopep8>=2.0.0

# CLI and Utilities
typer>=0.9.0
click>=8.1.0
rich>=13.5.0

# Logging and Monitoring
loguru>=0.7.0
structlog>=23.2.0

# Testing and Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.6.0

# Additional utilities
typing-extensions>=4.8.0